points changed from integer to double

+++ Point.java
@@ -1,28 +1,28 @@
 public class Point {
-    private int x,y;
+    private double x,y;
 
     public Point(){
         x=0; y=0;
     }
 
-    public Point(int x,int y){
+    public Point(double x,double y){
         this.x=x;
         this.y=y;
     }
 
-    public int getX(){ 
+    public double getX(){ 
         return  x;
     }
     
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
